--- pages/users/index.tsx
@@ -1,3 +1,0 @@
-export default function User() {
-  return <p>User table page</p>;
-}
